# src/swagger.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "MiniSense RESTful API Documentation",
        "contact": {
            "email": "alefs07@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:3333/",
            "description": "MiniSense RESTful API test"
        }
    ],
    "paths": {
        "/measurementUnit": {
            "post": {
                "summary": "Registering a measurement unit",
                "description": "Route responsible for registering a new unit of measure. No parameters are needed in the request, but you must assign a body in json format with a symbol and description of the new unit of measurement. If the symbol inserted in the body already exists, as a response you will receive a 400 response, but if the symbol does not exist you will receive a 200 response with a json with id, symbol and description.",
                "tags": ["Measurement Unit"],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MeasurementUnit"
                            },
                            "examples": {
                                "measurementUnit": {
                                    "value": {
                                        "symbol": "#$%",
                                        "description": "my measurement unit"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "400": {
                        "description": "Measurement unit already exists."
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "$ref": "#/components/schemas/MeasurementUnit"
                                },
                                "examples": {
                                    "measurementUnit": {
                                        "value": {
                                            "id": 1,
                                            "symbol": "#$%",
                                            "description": "my measurement unit"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "get": {
                "summary": "List of all measurement units",
                "description": "Route responsible for listing all measurement units. No parameters are needed in the request and no body. As a response you will receive a 200 response with a json with an array with all the measurement units registered or an empty array.",
                "tags": ["Measurement Unit"],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "$ref": "#/components/schemas/MeasurementUnit"
                                },
                                "examples": {
                                    "measurementUnit": {
                                        "value": [
                                            {
                                                "id": 1,
                                                "symbol": "ºC",
                                                "description": "Celsius"
                                            },
                                            {
                                                "id": 2,
                                                "symbol": "mg/m³C",
                                                "description": "Megagram per cubic metre"
                                            },
                                            {
                                                "id": 3,
                                                "symbol": "hPA",
                                                "description": "hectopasca"
                                            }
                                        ]
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/measurementUnit/{measurementUnitDescription}": {
            "get": {
                "summary": "Query a measurement unit by its description",
                "description": " Route responsible for querying a measurement unit for its description. You must pass the description of the measurement unit as a parameter. A bodysuit is not required. In response, if the description does not exist in the database you will receive a 400 response, but if the description exists you will receive a 200 response with a json with id, symbol and description.",
                "tags": ["Measurement Unit"],
                "parameters": [
                    {
                        "name": "measurementUnitDescription",
                        "in": "path",
                        "description": "Description of the measurement unit for consultation",
                        "required": true
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Measurement unit not found."
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "$ref": "#/components/schemas/MeasurementUnit"
                                },
                                "examples": {
                                    "measurementUnit": {
                                        "value": {
                                            "id": 1,
                                            "symbol": "#$%",
                                            "description": "my measurement unit"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/dataStream/{key}": {
            "post": {
                "summary": "Registration of a data stream in a sensor device",
                "description": "Route responsible for registering a new data stream to a sensory device through its key. The key of a sensory device must be passed as a parameter, and the body must be passed the stream label and the id of a unit of measurement. As an answer, if the device key is not found in the database or if the label already exists you will receive a 400 response, but if the key is found and the label does not exist you will receive a 200 response with a json with: id, key, label, unitId, deviceId and measurementCount.",
                "tags": ["Data Stream"],
                "parameters": [
                    {
                        "name": "key",
                        "in": "path",
                        "description": "Device key for query",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/DataStream"
                            },
                            "examples": {
                                "dataStream": {
                                    "value": {
                                        "label": "my label",
                                        "unitId": 1
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "400": {
                        "description": "Device or measurement unit not found."
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "$ref": "#/components/schemas/DataStream"
                                },
                                "examples": {
                                    "dataStream": {
                                        "value": {
                                            "id": 1,
                                            "key": "2af21e45-fa70-4ce2-a07f-cb62d4c8dd7c",
                                            "label": "my label",
                                            "unitId": 1,
                                            "deviceId": 1,
                                            "measurementCount": 0
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "get": {
                "summary": "Query a data stream by its key",
                "description": "Route responsible for querying a data stream by its key. The key of a data stream must be passed as a parameter. A bodysuit is not required. As a response, if the informed key is not found in the database you will receive a 400 response, but if the key is found you will receive a 200 response with a json with: id, key, label, unitId, deviceId, measurementsCount ans measurements array. All measurements will be reported.",
                "tags": ["Data Stream"],
                "parameters": [
                    {
                        "name": "key",
                        "in": "path",
                        "description": "Data stream key for query",
                        "required": true
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Data stream not found."
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "$ref": "#/components/schemas/DataStream"
                                },
                                "examples": {
                                    "dataStream": {
                                        "value": {
                                            "id": 1,
                                            "key": "19252280-6df3-4397-8138-38e74584ef1b",
                                            "label": "my label",
                                            "unitId": 1,
                                            "deviceId": 1,
                                            "measurementCount": 4,
                                            "measurements": [
                                                {
                                                    "timestamp": "2022-20-01 20:33:15",
                                                    "value": 259.43
                                                },
                                                {
                                                    "timestamp": "2022-20-01 20:33:30",
                                                    "value": 260.12
                                                },
                                                {
                                                    "timestamp": "2022-20-01 20:33:45",
                                                    "value": 259.96
                                                },
                                                {
                                                    "timestamp": "2022-20-01 20:34:00",
                                                    "value": 261.86
                                                }
                                            ]
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "MeasurementUnit": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "number"
                    },
                    "symbol": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    }
                }
                
            },
...
            "DataStream": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "number"
                    },
                    "key": {
                        "type": "string"
                    },
                    "label": {
                        "type": "string"
                    },
                    "enabled": {
                        "type": "boolean"
                    },
                    "measurementUnit_id": {
                        "type": "number"
                    },
                    "sensorDevice_id": {
                        "type": "number"
                    }
                }
            },
...
        }
    }
}